requirements.txt
google-genai
python-dotenv
sentence-transformers
faiss-cpu
numpy
PyPDF2
groq
schedule 
streamlit
schedule